Register repository in initial step
